refactor: moves Sidebar Provider to Sidebar file
